+++ dramatiq/brokers/rabbitmq.py
@@ -514,7 +514,11 @@ class _RabbitmqConsumer(Consumer):
             self._nack(method.delivery_tag)
             return None
 
-        rmq_message = _RabbitmqMessage(method.delivery_tag, message)
+        rmq_message = _RabbitmqMessage(
+            method.redelivered,
+            method.delivery_tag,
+            message,
+        )
         self.known_tags.add(method.delivery_tag)
         return rmq_message
 
@@ -545,7 +549,8 @@ class _RabbitmqConsumer(Consumer):
 
 
 class _RabbitmqMessage(MessageProxy):
-    def __init__(self, tag, message):
+    def __init__(self, redelivered, tag, message):
         super().__init__(message)
 
+        self.redelivered = redelivered
         self._tag = tag