+++ dramatiq/middleware/retries.py
@@ -92,6 +92,10 @@ class Retries(Middleware):
             return
 
         retries = message.options.setdefault("retries", 0)
+
+        message.options["retries"] += 1
+        message.options["traceback"] = traceback.format_exc(limit=30)
+
         max_retries = message.options.get("max_retries") or actor.options.get("max_retries", self.max_retries)
         retry_when = actor.options.get("retry_when", self.retry_when)
         if retry_when is not None and not retry_when(retries, exception) or \
@@ -99,9 +103,6 @@ class Retries(Middleware):
             self.logger.warning("Retries exceeded for message %r.", message.message_id)
             message.fail()
             return
-
-        message.options["retries"] += 1
-        message.options["traceback"] = traceback.format_exc(limit=30)
 
         if isinstance(exception, Retry) and exception.delay is not None:
             delay = exception.delay