+++ tests/middleware/test_threading.py
@@ -10,6 +10,7 @@ not_supported = threading.current_platform not in threading.supported_platforms
 
 
 @pytest.mark.skipif(not_supported, reason="Threading not supported on this platform.")
+@pytest.mark.skipif(threading.is_gevent_active(), reason="Thread exceptions not supported with gevent.")
 def test_raise_thread_exception():
     # Given that I have a database
     caught = []
@@ -36,15 +37,16 @@ def test_raise_thread_exception():
 
 
 @pytest.mark.skipif(not_supported, reason="Threading not supported on this platform.")
+@pytest.mark.skipif(threading.is_gevent_active(), reason="Thread exceptions not supported with gevent.")
 def test_raise_thread_exception_on_nonexistent_thread(caplog):
     # When an interrupt is raised on a nonexistent thread
-    threading.raise_thread_exception(-1, threading.Interrupt)
+    thread_id = 2 ** 31 - 1
+    threading.raise_thread_exception(thread_id, threading.Interrupt)
 
     # I expect a 'failed to set exception' critical message to be logged
+    expected_message = "Failed to set exception (Interrupt) in thread %d." % thread_id
     assert caplog.record_tuples == [
-        ("dramatiq.middleware.threading", logging.CRITICAL, (
-            "Failed to set exception (Interrupt) in thread -1."
-        )),
+        ("dramatiq.middleware.threading", logging.CRITICAL, expected_message),
     ]