+++ docs/source/changelog.rst
@@ -8,6 +8,51 @@ All notable changes to this project will be documented in this file.
 `Unreleased`_
 -------------
 
+Removed
+^^^^^^^
+
+* Dropped Python 3.5 support as it reached end-of-life
+
+
+`1.12.0`_ -- 2021-10-23
+-----------------------
+
+Added
+^^^^^
+
+* RabbitMQ messages now have a ``redelivered`` flag.  (`#405`_, `@nffdiogosilva`_)
+* Time limits now work under gevent.  (`#408`_, `@FinnLidbetter`_)
+* Shutdown notifications now work under gevent.  (`#426`_, `@FinnLidbetter`_)
+
+.. _#405: https://github.com/Bogdanp/dramatiq/pull/405
+.. _#408: https://github.com/Bogdanp/dramatiq/pull/408
+.. _#426: https://github.com/Bogdanp/dramatiq/pull/426
+.. _@FinnLidbetter: https://github.com/FinnLidbetter
+.. _@nffdiogosilva: https://github.com/nffdiogosilva
+
+Changed
+^^^^^^^
+
+* The ``watchdog`` library is no longer being pinned to a specific
+  version.  (`#428`_)
+* The redis broker now limits unpacks to half the size of the Lua
+  stack.  (`#433`_, `#434`_, `@ethervoid`_)
+
+.. _#428: https://github.com/Bogdanp/dramatiq/pull/428
+.. _#433: https://github.com/Bogdanp/dramatiq/issues/433
+.. _#434: https://github.com/Bogdanp/dramatiq/pull/434
+.. _@ethervoid: https://github.com/ethervoid
+
+Fixed
+^^^^^
+
+* Async exceptions now correctly set the thread id on Python 3.7 and
+  up.  (`#419`_, `#420`_, `@FinnLidbetter`_)
+
+.. _#419: https://github.com/Bogdanp/dramatiq/pull/419
+.. _#420: https://github.com/Bogdanp/dramatiq/pull/420
+.. _@FinnLidbetter: https://github.com/FinnLidbetter
+
 
 `1.11.0`_ -- 2021-05-22
 -----------------------
@@ -953,7 +998,8 @@ Changed
 * Capped prefetch counts to 65k.
 
 
-.. _Unreleased: https://github.com/Bogdanp/dramatiq/compare/v1.11.0...HEAD
+.. _Unreleased: https://github.com/Bogdanp/dramatiq/compare/v1.12.0...HEAD
+.. _1.12.0: https://github.com/Bogdanp/dramatiq/compare/v1.11.0...v1.12.0
 .. _1.11.0: https://github.com/Bogdanp/dramatiq/compare/v1.10.0...v1.11.0
 .. _1.10.0: https://github.com/Bogdanp/dramatiq/compare/v1.9.0...v1.10.0
 .. _1.9.0: https://github.com/Bogdanp/dramatiq/compare/v1.8.1...v1.9.0